feat : in progress

+++ src/main/java/com/etna/gpe/controller/OrganizationController.java
@@ -1,5 +1,6 @@
 package com.etna.gpe.controller;
 
+import com.etna.gpe.controller.customexception.ParametersNotFound;
 import com.etna.gpe.dto.OrganizationDto;
 import com.etna.gpe.model.Organization;
 import com.etna.gpe.service.OrganizationService;
@@ -25,19 +26,25 @@ public class OrganizationController {
     @PostMapping("/create_organization")
     @ResponseStatus(HttpStatus.CREATED)
     Organization createOrganization(@RequestBody OrganizationDto organizationDto) {
+    	if(organizationDto == null)
+    		throw new ParametersNotFound();
         return organizationService.createOrUpdateuOrganization(organizationDto);
     }
 
     @GetMapping("/get_organization")
     @ResponseStatus(HttpStatus.OK)
-    OrganizationDto getParticularByEmail(@RequestParam(value = "email") String email) {
+    OrganizationDto getOrganizationByEmail(@RequestParam(value = "email") String email) {
+    	if(email.isEmpty() || email == null)
+    		throw new ParametersNotFound();
         return organizationService.getOrganizationByEmail(email);
     }
 
-    //todo post a requestbodyn and not delete really but put boolean to true
+    //todo post a requestbody and not delete really but put boolean to true
     @PostMapping("delete_organization")
     @ResponseStatus(HttpStatus.RESET_CONTENT)
     void deleteparticular(@RequestParam(value = "email") String email) {
+    	if(email.isEmpty() || email == null)
+    		throw new ParametersNotFound();
         organizationService.deleteOrganization(email);
     }
