+++ src/main/java/com/etna/gpe/controller/ParticularController.java
@@ -1,5 +1,6 @@
 package com.etna.gpe.controller;
 
+import com.etna.gpe.controller.customexception.ParametersNotFound;
 import com.etna.gpe.dto.ParticularDto;
 import com.etna.gpe.model.Particular;
 import com.etna.gpe.service.ParticularService;
@@ -13,34 +14,38 @@ import java.util.List;
 @RequestMapping("/particular")
 public class ParticularController {
 
-    @Autowired
-    ParticularService particularService;
+	@Autowired
+	ParticularService particularService;
 
-    @GetMapping("/all_particular")
-    @ResponseStatus(HttpStatus.OK)
-    List<ParticularDto> getAllParticular() {
-        return particularService.getAllParticular();
-    }
+	@GetMapping("/all_particular")
+	@ResponseStatus(HttpStatus.OK)
+	List<ParticularDto> getAllParticular() {
+		return particularService.getAllParticular();
+	}
 
-    @PostMapping("/create_particular")
-    @ResponseStatus(HttpStatus.CREATED)
-    Particular createParticular(@RequestBody ParticularDto particularDto) {
-       return particularService.createOrUpdateuParticular(particularDto);
-    }
+	@PostMapping("/create_particular")
+	@ResponseStatus(HttpStatus.CREATED)
+	Particular createParticular(@RequestBody ParticularDto particularDto) {
+		if (particularDto == null)
+			throw new ParametersNotFound();
+		return particularService.createOrUpdateuParticular(particularDto);
+	}
 
-    @GetMapping("/get_particular")
-    @ResponseStatus(HttpStatus.OK)
-    ParticularDto getParticularByEmail(@RequestParam(value = "email") String email) {
-        return
-                particularService.getParticularByEmail(email);
-    }
-    
-  //todo post a requestbodyn and not delete really but put boolean to true
-    @PostMapping("/delete_particulier")
-    @ResponseStatus(HttpStatus.RESET_CONTENT)
-    void deleteparticular(@RequestParam(value = "email") String email) {
-        particularService.deleteParticular(email);
-    }
+	@GetMapping("/get_particular")
+	@ResponseStatus(HttpStatus.OK)
+	ParticularDto getParticularByEmail(@RequestParam(value = "email") String email) {
+		if (email == null || email.isEmpty())
+			throw new ParametersNotFound();
+		return particularService.getParticularByEmail(email);
+	}
 
+	// todo post a requestbodyn and not delete really but put boolean to true
+	@PostMapping("/delete_particulier")
+	@ResponseStatus(HttpStatus.RESET_CONTENT)
+	void deleteparticular(@RequestParam(value = "email") String email) {
+		if (email == null || email.isEmpty())
+			throw new ParametersNotFound();
+		particularService.deleteParticular(email);
+	}
 
 }
